src/requirements.txt
python-dotenv==1.0.1
Flask==3.0.2
Flask-SocketIO==5.3.6